+++ include/CeLux/conversion/cpu/BGRToRGB.hpp
@@ -2,8 +2,7 @@
 #pragma once
 
 #include "CPUConverter.hpp"
-#include "Frame.hpp"
-#include <type_traits>
+
 
 namespace celux
 {