+++ include/CeLux/conversion/cpu/CPUConverters.hpp
@@ -7,4 +7,6 @@
 #include <cpu/YUV420P10ToRGB48.hpp>
 #include <cpu/BGRToRGB.hpp>
 #include <cpu/RGBToRGB.hpp>
+#include <cpu/RGBAToRGB.hpp>
+#include <cpu/BGRAToRGB.hpp>
 #endif // CPU_CONVERTERS_HPP