+++ include/CeLux/conversion/gpu/cuda/P010LEToRGB.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "BaseConverter.hpp"
-#include "Frame.hpp"
+
 extern "C"
 {
     /**