+++ include/CeLux/python/VideoReader.hpp
@@ -3,10 +3,8 @@
 
 #include "Decoder.hpp" // Ensure this includes the Filter class
 #include "Factory.hpp"
-#include <memory> // For std::unique_ptr
 #include <pybind11/pybind11.h>
-#include <string> // For std::string
-#include <vector> // For std::vector
+
 
 namespace py = pybind11;