+++ tests/test1.py
@@ -65,11 +65,8 @@ def process_video_with_visualization(video_path):
     try:
         frame_count = 0
         start = time.time()
-        filter = celux.Scale(width = "1920", height = "1080", flags = "lanczos")
-        filters = [filter]
-        with celux.VideoReader(video_path, device = "cpu", num_threads = 16, filters = filters) as reader:
-            
-            input("Press Enter to start processing video...")
+
+        with celux.VideoReader(video_path, device = "cpu", num_threads = 16) as reader:
             for i, frame in enumerate(reader):
                 if frame_count == 0:
                     logging.info(